+++ src/Map.h
@@ -18,12 +18,12 @@ public:
 
 	void addQuadVertices(std::vector<sf::Vertex>& rvVertices) const;
 	void addTriangleVertices(std::vector<sf::Vertex>& rvVertices) const;
-	bool assignWorkplace(Worker& rWorker, const int x, const int y) const;
+	bool assignWorkplace(Worker& rWorker, const float x, const float y) const;
 
 private:
 	std::vector<std::unique_ptr<Structure>> m_vStructures;
 	std::vector<std::vector<std::unique_ptr<Tile>>> m_vTiles;
 	std::vector<std::unique_ptr<Worker>> m_vWorkers;
 
-	Workplace* getClosestFreeWorkplace(const int x, const int y) const;
+	Workplace* getClosestFreeWorkplace(const float x, const float y) const;
 };
